+++ src/InstanceBuilder.ts
@@ -9,9 +9,11 @@ import isObject from './isObject';
 export type Constructor<T> = new (...args: any[]) => T;
 
 export default class InstanceBuilder<T> extends Builder<T | null> {
+  loose: boolean = false;
+
   refClass: Constructor<T> | null = null;
 
-  constructor(refClass: Constructor<T> | null = null) {
+  constructor(refClass: Constructor<T> | null = null, loose: boolean = false) {
     super('instance', null);
 
     // Nullable by default
@@ -22,16 +24,19 @@ export default class InstanceBuilder<T> extends Builder<T | null> {
         this.invariant(typeof refClass === 'function', 'A class reference is required.');
       }
 
+      this.loose = loose;
       this.refClass = refClass;
       this.addCheck(this.checkInstance, refClass);
     }
   }
 
-  checkInstance(path: string, value: any, refClass: T | null) {
+  checkInstance(path: string, value: T, refClass: T | null) {
     if (__DEV__) {
       if (refClass) {
         this.invariant(
-          typeof refClass === 'function' && value instanceof refClass,
+          typeof refClass === 'function' &&
+            (value instanceof refClass ||
+              (this.loose && isObject(value) && value.constructor.name === refClass.name)),
           `Must be an instance of "${this.typeAlias()}".`,
           path,
         );
@@ -55,14 +60,17 @@ export default class InstanceBuilder<T> extends Builder<T | null> {
   }
 }
 
-export function instance<T>(refClass: Constructor<T> | null = null): InstanceBuilder<T> {
-  return new InstanceBuilder(refClass);
+export function instance<T = Function>(
+  refClass: Constructor<T> | null = null,
+  loose?: boolean,
+) /* infer */ {
+  return new InstanceBuilder<T>(refClass, loose);
 }
 
-export function regex(): InstanceBuilder<RegExp> {
+export function regex() /* infer */ {
   return instance(RegExp);
 }
 
-export function date(): InstanceBuilder<Date> {
+export function date() /* infer */ {
   return instance(Date);
 }