+++ src/NumberBuilder.ts
@@ -5,13 +5,13 @@
 
 import Builder from './Builder';
 
-function isNumber(value: any): boolean {
+function isNumber(value: any): value is number {
   return typeof value === 'number';
 }
 
-export default class NumberBuilder extends Builder<number | null> {
-  constructor(defaultValue: number | null = 0) {
-    super('number', defaultValue);
+export default class NumberBuilder<T extends number = number> extends Builder<T> {
+  constructor(defaultValue?: T) {
+    super('number', defaultValue || (0 as T));
   }
 
   between(min: number, max: number, inclusive: boolean = false): this {
@@ -25,7 +25,7 @@ export default class NumberBuilder extends Builder<number | null> {
     return this.addCheck(this.checkBetween, min, max, inclusive);
   }
 
-  checkBetween(path: string, value: any, min: number, max: number, inclusive: boolean = false) {
+  checkBetween(path: string, value: T, min: number, max: number, inclusive: boolean = false) {
     if (__DEV__) {
       this.invariant(
         isNumber(value) && (inclusive ? value >= min && value <= max : value > min && value < max),
@@ -47,7 +47,7 @@ export default class NumberBuilder extends Builder<number | null> {
     return this.gt(min, true);
   }
 
-  checkGreaterThan(path: string, value: any, min: number, inclusive: boolean = false) {
+  checkGreaterThan(path: string, value: T, min: number, inclusive: boolean = false) {
     if (__DEV__) {
       if (inclusive) {
         this.invariant(
@@ -73,7 +73,7 @@ export default class NumberBuilder extends Builder<number | null> {
     return this.lt(max, true);
   }
 
-  checkLessThan(path: string, value: any, max: number, inclusive: boolean = false) {
+  checkLessThan(path: string, value: T, max: number, inclusive: boolean = false) {
     if (__DEV__) {
       if (inclusive) {
         this.invariant(
@@ -87,7 +87,7 @@ export default class NumberBuilder extends Builder<number | null> {
     }
   }
 
-  oneOf(list: number[]): this {
+  oneOf<U extends number>(list: U[]) /* refine */ {
     if (__DEV__) {
       this.invariant(
         Array.isArray(list) && list.length > 0 && list.every(item => isNumber(item)),
@@ -95,16 +95,18 @@ export default class NumberBuilder extends Builder<number | null> {
       );
     }
 
-    return this.addCheck(this.checkOneOf, list);
+    this.addCheck(this.checkOneOf, list);
+
+    return (this as any) as NumberBuilder<U>;
   }
 
-  checkOneOf(path: string, value: any, list: number[]) {
+  checkOneOf(path: string, value: T, list: T[]) {
     if (__DEV__) {
       this.invariant(list.indexOf(value) >= 0, `Number must be one of: ${list.join(', ')}`, path);
     }
   }
 }
 
-export function number(defaultValue: number | null = 0): NumberBuilder {
-  return new NumberBuilder(defaultValue);
+export function number<T extends number = number>(defaultValue?: number) /* infer */ {
+  return new NumberBuilder<T>(defaultValue as T);
 }