+++ src/index.ts
@@ -5,20 +5,37 @@
 
 import optimal from './optimal';
 import Builder, { bool, custom, func } from './Builder';
-import CollectionBuilder, { array, object } from './CollectionBuilder';
+import ArrayBuilder, { array } from './ArrayBuilder';
 import InstanceBuilder, { instance, date, regex } from './InstanceBuilder';
 import NumberBuilder, { number } from './NumberBuilder';
+import ObjectBuilder, { object } from './ObjectBuilder';
 import ShapeBuilder, { shape } from './ShapeBuilder';
 import StringBuilder, { string } from './StringBuilder';
 import UnionBuilder, { union } from './UnionBuilder';
+
+export const predicates = {
+  array,
+  bool,
+  custom,
+  date,
+  func,
+  instance,
+  number,
+  object,
+  regex,
+  shape,
+  string,
+  union,
+};
 
 export { array, bool, custom, date, func, instance, number, object, regex, shape, string, union };
 
 export {
   Builder,
-  CollectionBuilder,
+  ArrayBuilder,
   InstanceBuilder,
   NumberBuilder,
+  ObjectBuilder,
   ShapeBuilder,
   StringBuilder,
   UnionBuilder,