+++ src/isObject.ts
@@ -6,6 +6,6 @@
 /**
  * Return true if the value is a plain object.
  */
-export default function isObject(value: any): boolean {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
+export default function isObject(value: any): value is object {
+  return !!value && typeof value === 'object' && !Array.isArray(value);
 }