+++ tests/NumberBuilder.test.ts
@@ -16,6 +16,7 @@ describe('NumberBuilder', () => {
   describe('runChecks()', () => {
     it('errors if a non-number value is used', () => {
       expect(() => {
+        // @ts-ignore Testing wrong type
         number().runChecks('key', 'foo', {});
       }).toThrowErrorMatchingSnapshot();
     });
@@ -24,14 +25,14 @@ describe('NumberBuilder', () => {
   describe('between()', () => {
     it('errors if min is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.between('foo', 5);
       }).toThrowErrorMatchingSnapshot();
     });
 
     it('errors if max is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.between(0, 'foo');
       }).toThrowErrorMatchingSnapshot();
     });
@@ -49,6 +50,7 @@ describe('NumberBuilder', () => {
   describe('checkBetween()', () => {
     it('errors if value is not a number', () => {
       expect(() => {
+        // @ts-ignore Testing wrong type
         builder.checkBetween('key', 'foo', 0, 5);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -81,7 +83,7 @@ describe('NumberBuilder', () => {
   describe('gt()', () => {
     it('errors if min is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.gt('foo');
       }).toThrowErrorMatchingSnapshot();
     });
@@ -99,7 +101,7 @@ describe('NumberBuilder', () => {
   describe('gte()', () => {
     it('errors if min is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.gte('foo');
       }).toThrowErrorMatchingSnapshot();
     });
@@ -117,6 +119,7 @@ describe('NumberBuilder', () => {
   describe('checkGreaterThan()', () => {
     it('errors if value is not a number', () => {
       expect(() => {
+        // @ts-ignore Testing wrong type
         builder.checkGreaterThan('key', 'foo', 5);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -149,7 +152,7 @@ describe('NumberBuilder', () => {
   describe('lt()', () => {
     it('errors if max is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.lt('foo');
       }).toThrowErrorMatchingSnapshot();
     });
@@ -167,7 +170,7 @@ describe('NumberBuilder', () => {
   describe('lte()', () => {
     it('errors if max is a not a number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.lte('foo');
       }).toThrowErrorMatchingSnapshot();
     });
@@ -185,6 +188,7 @@ describe('NumberBuilder', () => {
   describe('checkLessThan()', () => {
     it('errors if value is not a number', () => {
       expect(() => {
+        // @ts-ignore Testing wrong type
         builder.checkLessThan('key', 'foo', 5);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -217,7 +221,7 @@ describe('NumberBuilder', () => {
   describe('oneOf()', () => {
     it('errors if not an array', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.oneOf(123);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -230,7 +234,7 @@ describe('NumberBuilder', () => {
 
     it('errors if array contains a non-number', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.oneOf(['foo', 123]);
       }).toThrowErrorMatchingSnapshot();
     });